release 2

+++ studentManagementPortal/src/main/java/com/student/management/studentManagementPortal/dao/StudentDao.java
@@ -3,5 +3,5 @@ package com.student.management.studentManagementPortal.dao;
 import com.student.management.studentManagementPortal.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentDao extends JpaRepository<Student,Long> {
+public interface StudentDao extends JpaRepository<Student, Integer> {
 }
